refactor(example): extract native E2E scenario modules
